markup terms in Key-Terms.md as linkable definitions
Though GFM will strip `<dfn>` and thus the fragment identifiers, the jenkins markdown processor, kramdown, should preserve ’em. One could optionally instead mark these up with `<a id="term-foo"><dfn>...</dfn></a>` so that the fragment ID links work in Github’s own views in addition to the Github Pages artifacts, but I figured it’d be a pain to do that and it strikes me as decidedly more important fer the artifact’s markup be clear and coherent.